requirements.txt
langchain
langchain-core
langchain_community
sentence-transformers
umap-learn
langchain_groq
